+++ src/stactools/hotosm/stac.py
@@ -28,15 +28,12 @@ from stactools.hotosm.constants import (
     COLLECTION_DESCRIPTION,
     COLLECTION_ID,
     COLLECTION_TITLE,
+    OAM_EXTENSION_DEFAULT_VERSION,
+    OAM_EXTENSION_SCHEMA_URI_PATTERN,
 )
 from stactools.hotosm.exceptions import AssetNotFoundError
 from stactools.hotosm.oam_metadata import OamMetadata
 from stactools.hotosm.stac_common import add_alternate_assets
-
-OAM_EXT_VERSION = "v0.1.0"
-OAM_EXT_SCHEMA = (
-    f"https://hotosm.github.io/stactools-hotosm/oam/{OAM_EXT_VERSION}/schema.json"
-)
 
 
 def create_collection() -> Collection:
@@ -73,12 +70,9 @@ def create_collection() -> Collection:
     )
 
     collection.item_assets = {
-        "image": ItemAssetDefinition.create(
+        "visual": ItemAssetDefinition.create(
             title="Visual image",
-            description=(
-                "Visual imagery data acquired from satellite or unmanned aerial "
-                "vehicle (UAV)"
-            ),
+            description=("Imagery data formatted for visualization (usually RGB)"),
             media_type=MediaType.COG,
             roles=["data"],
         ),
@@ -96,7 +90,7 @@ def create_collection() -> Collection:
         {
             "visual": Render.create(
                 assets=[
-                    "image",
+                    "visual",
                 ],
                 title="Visual image",
             )
@@ -164,7 +158,7 @@ def create_item(oam_metadata: OamMetadata) -> Item:
         item.properties["created"] = datetime_to_str(oam_metadata.uploaded_at)
 
     item.add_asset(
-        "image",
+        "visual",
         Asset(
             href=oam_metadata.image_url,
             title=oam_metadata.title,
@@ -174,7 +168,7 @@ def create_item(oam_metadata: OamMetadata) -> Item:
     )
 
     item.ext.add("file")
-    file_ext = FileExtension.ext(item.assets["image"])
+    file_ext = FileExtension.ext(item.assets["visual"])
     file_ext.apply(size=oam_metadata.image_file_size)
 
     item.add_asset(
@@ -197,8 +191,12 @@ def create_item(oam_metadata: OamMetadata) -> Item:
         ),
     )
 
-    _add_projection_extension(item, ["image"])
+    _add_projection_extension(item, ["visual"])
     add_alternate_assets(item)
+
+    item.stac_extensions.append(
+        OAM_EXTENSION_SCHEMA_URI_PATTERN.format(version=OAM_EXTENSION_DEFAULT_VERSION)
+    )
 
     item.validate()